requirements.txt
langchain 
langchain_community 
PyPDF2 
openai==0.28 
tiktoken 
faiss-cpu
streamlit
python-dotenv